@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:jordan/extras/statics.dart';
 
 class CalendariumWidget extends StatelessWidget {
-  const CalendariumWidget({Key key}) : super(key: key);
+  const CalendariumWidget({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
